Moved caches into a subdir/own namespace
